Homework 8
I noticed the runByBin function had been removed from the previous
commit some time between testing and commit, as it is a fatal flaw and
cannot run without it and all tests completed.

Sorry for the oversight.

This commit also includes error handling for each of the functions in
explore.
the convention was to run a series of tests and create a list of the
test results.

Then use this list to return some explanation of the error and the
variable which causes the error.  Some of these errors will stop the
program and others will merely print the error message to the console
informing the user of the error.